feat: done forms stylization
