+++ src/components/ControlledForm/Elements/SelectControlled.tsx
@@ -1,10 +1,11 @@
 import type { UseFormRegisterReturn } from 'react-hook-form';
+import styles from '../../../styles/form.module.css';
 
 export function SelectControlled({ register }: InputElementProps) {
   const countries = ['Russia', 'USA', 'Belarus'];
 
   return (
-    <>
+    <div className={styles.inputWrapper}>
       <legend>{'Choose your country'}:</legend>
       <select {...register}>
         {countries.map((country) => (
@@ -13,7 +14,7 @@ export function SelectControlled({ register }: InputElementProps) {
           </option>
         ))}
       </select>
-    </>
+    </div>
   );
 }
