+++ src/components/UncontrolledForm/Elements/SelectUncontrolled.tsx
@@ -3,7 +3,7 @@ export function SelectUncontrolled() {
 
   return (
     <>
-      <label htmlFor="city-select">Choose your country</label>
+      <label htmlFor="city-select">Choose your country: </label>
       <select name="city" id="city-select">
         {countries.map((country) => (
           <option key={country} value={country}>
